Add new project: Magic Sort
