+++ Function Pointer/cmake-build-debug/Testing/Temporary/LastTest-DESKTOP-T5M6G1B.log
@@ -1,0 +1,3 @@
+Start testing: May 19 13:30 SE Asia Standard Time
+----------------------------------------------------------
+End testing: May 19 13:30 SE Asia Standard Time
