+++ Hello World/cmake-build-debug/CMakeFiles/Hello_World.dir/manifest.rc
@@ -1,0 +1,2 @@
+#pragma code_page(65001)
+1 /* CREATEPROCESS_MANIFEST_RESOURCE_ID */ 24 /* RT_MANIFEST */ "C:/Users/Trong_Phuc/OneDrive_VNU-HCMUS/Documents/JetBrains_Tools/Clion/C projects/Hello World/cmake-build-debug/CMakeFiles/Hello_World.dir/embed.manifest"
