+++ Hello World/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: May 25 16:14 SE Asia Standard Time
+Start testing: May 30 23:57 SE Asia Standard Time
 ----------------------------------------------------------
-End testing: May 25 16:14 SE Asia Standard Time
+End testing: May 30 23:57 SE Asia Standard Time
